[xfa] integer overflow vulnerability on RelocateTableRowCells.
An integer overflow flaw which could lead to an out-of-bounds-read
was discovered in RelocateTableRowCells (Issue 996770).

Bug: chromium:996770

+++ xfa/fxfa/layout/cxfa_contentlayoutprocessor.cpp
@@ -222,8 +222,8 @@ void RelocateTableRowCells(const RetainPtr<CXFA_ContentLayoutItem>& pLayoutRow,
     int32_t nColSpan = nOriginalColSpan;
     float fColSpanWidth = 0;
     if (nColSpan == -1 ||
-        nCurrentColIdx + nColSpan >
-            pdfium::CollectionSize<int32_t>(rgSpecifiedColumnWidths)) {
+        (uint32_t)nCurrentColIdx + (uint32_t)nColSpan >
+            pdfium::CollectionSize<uint32_t>(rgSpecifiedColumnWidths)) {
       nColSpan = pdfium::CollectionSize<int32_t>(rgSpecifiedColumnWidths) -
                  nCurrentColIdx;
     }
